Exit if reconmfiguration fails
When going from 2 to 1 zookeeper servers (after going from 1 to 2 nodes
because a node is about to be retired), reconfiguration will fail
if the second node is removed before reconfiguration has happened.
Reconfiguration will never work in these cases, since 2 nodes is
needed to have ZooKeepere quorum. Just exit if this happens, new
config will be used in this case and ZooKeeper should work fine again.