Take dead elements into account (and subtract them) when calculating how many arrays to allocate in a datastore buffer.
This avoids a problem were allocated memory can accumulate over time in components using an ArrayStore.
If all documents in an array attribute vector changes from one value class to another,
all elements in the buffers of the previous value class are marked dead.
Those buffers will eventually be compacted.
Without this fix the wanted size of the resulting compacted buffer is calculated too high,
and we allocate memory we are not going to use. If we move to yet another value class later,
the same problem occurs again and more memory is allocated.