Increase time to keep unused file references
In case of network issues, failed deployments etc. there might be a need
to serve unused file references for some time (think an upgrade that takes
a long time and a node that comes back to life on an old version).

Also change to use minutes as resolution to make it easier to write tests.