Remove notion of node-specific reliability from Java distribution code
Not used, and wasn't algorithmically in sync with the C++ code anyway.
Also add guard to avoid emitting invalid node indices for storage nodes
if the number of configured nodes is lower than the replication factor.
Looks like this particular code path is only called by cross-language
conformance tests, so hasn't been a problem in practice.