Simplify calculating delay after config failures
When there are several subscribers to a config and 1 JRTConfiguRequester
the counter for failures might quickly increase so backoff based
on these numbers will not be useful. Simplify by using booleans
for failures instead. Max delay for configured subscriptions will
now be 30 seconds instead of 150 seconds (which is way too much)