Use local sessions fromm file system when deleting expired sessions
Earlier we loaded all local sessions at startup time and added to
a cache, that was used when deleting expired sessions.

We don't load local sessions at startup that anymore, but sessions are
still added to cache as they are created. Local sessions that were
leftover for some other reason (crash, failure in logic, whatnot) should
still be deleted, which is fixed with this.