Enable group-suspension with permanent suspension
This will enable the same per-cluster policy for "permanently suspending" a
host (i.e. removing it from an application, typically because it is
wantToRetire and retired), as the policy for suspending a host (typically for
upgrade).  The major difference is that the new policy allows for permanently
remove >1 content node at a time, if they are within the same group.

Guarded by a new flag group-suspension-in-permanent-suspend, enabled by
default.  The old and unused flag group-suspension is removed.