Prepare all applications before activating when bootstrapping config server
When upgrading to a new version when bootstrapping config server:

If we fail to prepare an application it might be because there is something
wrong with the release. In that case we don't want to activate any application.
This change makes sure we prepare all applications successfully before we
activate any application..