Try to generalise from the toSlime() method
There are multiple types of endpoints and routing methods. Here we
try to use the same logic as the toSlime() method to generate the
status JSON from the instance/endpoint POV.

Not sure if the JSON format makes sense or if are in a position where
we can get multiple contradictory statuses per endpoint+zone.