use saturated sum as hit estimate for OR
The upper bound for the hit estimate is the docid limit. The lower
bound for the docid limit is the maximum child estimate.

In the future we probably want a solution where everyone knows the
docid limit from the start, which might also require us to have a
separate way of estimating the document frequency if we want to avoid
ranking score regression.