Remove explicit syncing of all write executors when handling DeleteBucket operations.
This really hurts the performance of the feed pipeline (e.g. during merging buckets),
as all executor queues are drained and then the executors need time to rebuild momentum afterwards.
This is also not needed from a data consistency point of view, as the DeleteBucket operation is
ack'ed right after it has been written to the TLS.