Make the bill end date inclusive instead of exclusive
Before the end date for the billing was used to calculate startOfDay
to generate the timestamp we actually use.  Now we are changing it
to .plusDays(1).startOfDay to make it so UI will effectively use
the end of today as the timestamp.

Also added some more implementations to ManualClock to get some
tests to work.