Replace session timeout with explicit shutdown
Session timeout causes message bus to reply with timeouts when timeout passes.
This works poorly with visitors whose document put acks are delayed until the
network layer consumes the documents, which may take longer than the remaining
session timeout, which is used as message timeout.
Keeping the message timeout fixed, and doing a manual abort of the session instead,
when the specified timeout has occurred, almost eliminates the problem.
Additionally, acking all outstanding documents upon abortion makes the visitors
return in a timely manner, and should take care of the rest of the problem.