Don't trigger implicit ContentPolicy random send fallback on expected transient errors
The `ContentPolicy` has a failure handling policy where more than _n_ error
replies (a small number in practice) will trigger an implicit random send instead
of using the cached cluster state. This is to force rediscovery of the actual
cluster state, and is useful if a node is bad but we're not sending feed to enough
other nodes to figure it out from them.

However, certain error codes may be used frequently by the content layer for
purposes that do _not_ indicate that a change in cluster state may have happened,
and should therefore not be counted as errors that may indicate a bad node:
 * `ERROR_TEST_AND_SET_CONDITION_FAILED`: may happen for any mutating operation
   that has an associated TaS condition. Technically an `APP_FATAL_ERROR` since
   resending doesn't make sense.
 * `ERROR_BUSY`: may happen for concurrent mutations and if distributors are
   in the process of changing bucket ownership and the grace period hasn't
   passed yet. Also sent if queues are full and client policy should back off
   a bit. None of these are errors as per se.