Let CommunicationManager swallow any errant internal reply messages
It's possible for internal commands to have replies auto-generated
if they're being evicted/aborted from persistence queue structures.
Most of these replies are intercepted by higher-level links in the
storage chain, but commands such as `RunTaskCommand` are actually
initiated by the persistence _provider_ and not a higher level component,
and are therefore not caught explicitly anywhere.

Let CommunicationManager signal that internal replies are handled,
even if the handling of these is just to ignore them entirely. This
prevents the replies from being spuriously warning-logged as "unhandled
message on top of call chain".