Avoid clearing config response queue
There have been races due to the fact that we used to clear the queue
when receiving a response, thus missing some of the responses. This
stops clearing the queue and handles several items on the queue by
polling until the queue is empty when a new item is found on the queue.