Sample and track the current transient memory usage across all document dbs.
Previously, the transient memory usage was the theoretical worst case needed during loading of attributes.
Now, we change this to being a snapshot of the total memory used by all memory indexes.
In a simular fashion we plan to track the transient disk usage of all ongoing fusion disk indexes.

Eventually, the transient resource usage will be subtracted from the absolute resource usage,
before this is reported to the cluster controller in order to decide whether to block feeding.
By doing this the transient resource usage is covered by the headroom on the content node,
instead of leading to feed blocked due to natural fluctuations.