Update C++ DynamicThrottlePolicy to match Java impl semantics
This is an attempt at making the C++ implementation 1-1 with the
Java implementation, the latter considered the source of truth for
this matter. Unit tests have been updated to match their Java
equivalents as well.

This task was approached much in the same way a novice wizard would
transcribe arcane runes that may possible blow their hands off;
carefully and with more pragmatic respect than understanding.