Avoid assertion failure on aborted node startup
If a node fails to completely start up due to e.g. failures
fetching bootstrap config, it's possible for the storage component
chain to have been created but not yet completely initialized.
Attempts to gracefully abort startup will still invoke the component
`onClose()` methods, at which point the components must be able to
handle a partially initialized state. Let `ModifiedBucketChecker`
handle that its worker thread was not created when closing, as
this will happen when `onOpen()` is never called.