Deprecate remnants of two-phase visiting APIs and usage
This has not been relevant since "orderdoc" was a thing, and it
was never really a thing in the first place.

Unfortunately, due to an oversight in how the backend propagates visitor
statistics, the internal usage of 2nd pass statistics cannot be fully
removed before Vespa 8 (where the backend is known to not set the
deprecated statistics fields).