Avoid visitor reply processing starvation and runaway throttling
We both send requests and process replies in the context of a dedicated task executor pool.
However, MessageBus sending and reply receiving happens in the context of entirely
separate threads. If the backend responds very quickly to visitor requests (such as
if buckets are empty), this can leave us in the following awkward position:

  1. Replies arrive from backend, open up the throttle window, reply handling
     task gets pushed onto executor queue (but not yet executed).
  2. Send loop below continuously get a free send slot, keeps sending visitors
     and filling up the set of pending buckets in the progress token.
  3. Since visitor session is busy-looping in the send task, reply processing is
     consequently entirely starved until the MessageBus throttle window is bursting
     at the seams. This can effectively nullify the effects of the throttling policy,
     especially if it's dynamic. But a static throttle policy with a sufficiently
     high max window size will also potentially cause a runaway visitor train since
     the active window size keeps getting decreased by backend replies.

To get around this, we explicitly check for concurrently scheduled message handling
tasks from the transport layer, breaking the loop if at least one handler has been
scheduled. This also has the (positive) effect of draining all reply tasks before we
start sending more work downstream.

Since visitor session progress is edge-triggered and progresses exclusively by sending
new visitors in reply handling tasks, it's critical that we never end up in a situation
where we have no pending CreateVisitors (or scheduled tasks), or we risk effectively
hanging the session. We must therefore be very careful that we only exit the send loop
if we _know_ we have at least one pending task enqueued that will ensure session progress.

This is a subtle thread interaction issue and cannot readily be unit tested, but
manual testing has confirmed both the underlying issue and that the throttling policy
does not run amok after the fix has been applied.