Make B-tree bucket database values atomic to ensure well-defined access
Existing implementation already used explicit acquire/release fences to
ensure visibility from writer to concurrent readers, but the values written/read
were not of an atomic type and thus _technically_ considered a data race.

This commit adds an AtomicValueWrapper wrapper to vespalib which looks and acts
much like the existing AtomicEntryRef, but for primitive types that are not
related to EntryRefs. The bucket DB B-tree primitive u64 value type is replaced
with an atomic wrapper and explicit memory fences are replaced with release stores
and acquire loads on the atomic values themselves to ensure they form correct
pairs between writer and readers.