fix tsan warnings related to match phase limiting
This fixes the tsan warnings. The consequences of 'wrong' would
probably be incorrect coverage result or similar. However, the fact
that the race warning indicates that we read the was_used/estimate
BEFORE writing them could indicate that not all threads agree on
whether we want to limit the search or not (they should). Not sure if
this is a real issue or how serious it is.