Let empty entry BufferType sentinel be static instead of global
Avoids issue where a global BufferType with a transitive Alloc instance
binds to a null reference allocator since the global allocator instance
has not yet been initialized as part of global constructor invocation.

Manually hoist empty sentinel ref outside of loops since it might not
be obvious to the compiler that the same object is reused over and over.