Choose node resources with a matching host flavor when exclusive
When using a custom cloud account (always exclusive) we cannot choose a too
small flavor because there may not be any matching host flavor. This currently
works in our own zones because there is always a shared host that can be used
for admin nodes (feature flag is set in all zones) and there is no way to set
exclusivity requirement for those clusters.