Add support for cross-thread stack dumping to vespalib
There's no existing API for conveniently getting the stack trace of any other thread
than your own, so we have to turn things around a bit and make the target thread
dump its own stack in a way that we can then safely access and return it.

We allow for hooking SIGUSR2 at process startup time and use this as our dedicated
stack dump signal internally. The process of stack dumping is then basically:

 1. Caller sends SIGUSR2 to the target thread. This happens within a global mutex
    that ensures no other callers can attempt to dump stack at the same time. This
    is because we use a single, globally shared state between caller and target
    (the same signal handler function is used by all threads).
 2. Caller acquire-polls (with 1ms sleep) for target thread signal handler completion.
    Fancy technologies such as mutexes and condition variables are not safe to use
    from within a signal handler and therefore cannot be used.
 3. Target thread suddenly finds itself in Narnia (signal handler). Since the
    signal is the magical SIGUSR2, it proceeds to dump its stack frame addresses in
    a shared buffer. It then toggles completion, with release semantics, before
    returning to its regularly scheduled programming.
 4. Caller exits poll-loop and assembles a complete stack trace from the frame
    addresses in the shared buffer, demangled and shiny.

Signal-safe stack frame dumping currently uses Boost.Stacktrace, which is a
header-only dependency and appears to use libunwind under the hood.