Always use Vespa version in config request when serving config
This change means that nodes on a version that there exists no config
model for will not get new config, they will get an error instead.
This means that there will be no longer be possible to get incompatible
config for nodes on another version than one known by the config server.
https://github.com/vespa-engine/vespa/issues/23223