Simplify and generalize
Let embedder rewrites simply always change <x path="y" url="z">
to <xPath>y</xPath> <xUrl>z</xUrl>, as well as translating
the id attirebute to the corresponding path if on hosted.

This means that the current built-in embedder accepting
"vocab" and "model" is broken as these names are different
from the names in the config model, but this isn't documented yet
so should be okay.

The effect of this is that our built-in models can be
used in any embedder and config, and the embedder syntax
can be used in application specific embedders.