Use Java standard map instead of GSON internal one
LinkedHashTreeMap is an internal class never intended for use outside of GSON.
It was removed in https://github.com/google/gson/pull/1992 (version 2.9.0) and
its use in this project prevents the GSON version from being bumped.

The standard LinkedHashMap is a perfect drop-in replacement as it
also preserves insertion order.